# Improve performance of writer.