--- a/R/tags.R
+++ b/R/tags.R
@@ -421,7 +421,7 @@ tagWrite <- function(tag, textWriter, indent=0, eol = "\n") {
   }
 
   # write tag name
-  textWriter$write(paste8("<", tag$name, sep=""))
+  textWriter$write(paste0("<", tag$name))
 
   # Convert all attribs to chars explicitly; prevents us from messing up factors
   attribs <- lapply(tag$attribs, as.character)
@@ -447,10 +447,10 @@ tagWrite <- function(tag, textWriter, indent=0, eol = "\n") {
       if (is.logical(attribValue))
         attribValue <- tolower(attribValue)
       text <- htmlEscape(attribValue, attribute=TRUE)
-      textWriter$write(paste8(" ", attrib,"=\"", text, "\"", sep=""))
+      textWriter$write(paste0(" ", attrib,"=\"", text, "\""))
     }
     else {
-      textWriter$write(paste8(" ", attrib, sep=""))
+      textWriter$write(paste0(" ", attrib))
     }
   }
 
@@ -461,8 +461,7 @@ tagWrite <- function(tag, textWriter, indent=0, eol = "\n") {
 
     # special case for a single child text node (skip newlines and indentation)
     if ((length(children) == 1) && is.character(children[[1]]) ) {
-      textWriter$write(paste8(normalizeText(children[[1]]), "</", tag$name, ">",
-        sep=""))
+      textWriter$write(paste0(normalizeText(children[[1]]), "</", tag$name, ">"))
     }
     else {
       if ("after-begin" %in% .noWS || "inside" %in% .noWS) {
@@ -475,7 +474,7 @@ tagWrite <- function(tag, textWriter, indent=0, eol = "\n") {
       if ("before-end" %in% .noWS || "inside" %in% .noWS) {
         textWriter$eatWS()
       }
-      textWriter$write(paste8("</", tag$name, ">", sep=""))
+      textWriter$write(paste0("</", tag$name, ">"))
     }
   }
   else {
@@ -487,7 +486,7 @@ tagWrite <- function(tag, textWriter, indent=0, eol = "\n") {
       textWriter$write("/>")
     }
     else {
-      textWriter$write(paste8("></", tag$name, ">", sep=""))
+      textWriter$write(paste0("></", tag$name, ">"))
     }
   }
   if ("after" %in% .noWS || "outside" %in% .noWS) {