- Use a buffer to cache writes instead of writing and seeking on a binary unliked connection. See https://wch.github.io/string_builder/ for impact on performance.
 - collapse the textwriter and wstextwriter into a single entity to spare all the extra dispatch
 - eliminate R6 and just use closures to spare ourselves the overhead
 - migrate from paste8 to paste0 since write UTF8-encodes all inputs.

TODO: 

 - [x] confirm that paste0 won't mess up any encoding

## Performance Comparison

### on master (eea92)

```
> size <- 1000
> l <- tagList(lapply(1:size, function(x){htmltools::tags$option(value=x)}))
> bench::mark(min_time=10,{
+   s <- as.character(l)
+ })
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 { s <- as.character(l) } 118ms  162ms      6.24     623KB     18.7    63   189
# … with 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#   time <list>, gc <list>
```

### on joe/feature/nows (b4dcc)

```
> size <- 1000
> l <- tagList(lapply(1:size, function(x){htmltools::tags$option(value=x)}))
> bench::mark(min_time=10,{
+   s <- as.character(l)
+ })
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 { s <- as.character(l) } 219ms  257ms      3.89     499KB     7.08    39    71
# … with 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#   time <list>, gc <list>
```

### on jeff/feature/fast-nows(a038)

```
> size <- 1000
> l <- tagList(lapply(1:size, function(x){htmltools::tags$option(value=x)}))
> bench::mark(min_time=10,{
+   s <- as.character(l)
+ })
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 { s <- as.character(l) } 130ms  163ms      5.88     895KB     16.7    59   167
# … with 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#   time <list>, gc <list>
```