@trestletech 2019-07-11
Incorporated feedback! I left the handling of multiple arguments in `write` as a TODO. We have some more complex interactions around `paste8` at the moment, but it seems like we are doubly encoding into utf8, so it might be wise to collapse these behaviors. It just didn't look like it was going to be free, but I may poke at it.