@trestletech 2019-07-11
With the paste0 substitution in dce7d11, I'm now seeing a real performance boost over master.

```
> size <- 1000
> l <- tagList(lapply(1:size, function(x){htmltools::tags$option(value=x)}))
> bench::mark(min_time=10,{
+   s <- as.character(l)
+ })
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 { s <- as.character(l) } 112ms  124ms      7.83     792KB     135.     4    69
# … with 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#   time <list>, gc <list>
```

But @wch pointed out that we need to confirm that the `paste0() + write()` combination still produces sane encoding.