@trestletech 2019-07-23
Feedback incorporated

As of latest commit (e0262), we're about 10% faster than master.

```
# A tibble: 1 x 13
  expression                 min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result
  <bch:expr>               <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>
1 { s <- as.character(l) } 112ms  145ms      6.97     796KB     7.57    70    76        10s <chr …
# … with 3 more variables: memory <list>, time <list>, gc <list>
```