@trestletech 2019-07-23
Passing along some prior encoding research from Winston regarding whether or not we could further optimize `paste8`:

> Oh, wait, I just tried something else that has the bad behavior. Earlier, I said that `paste(x,y)` would convert both to UTF-8 before pasting, if one of them was UTF-8. But if you do `paste(c(x,y), collapse="")`, then it doesn't work that way.

```
Sys.setlocale(, "Chinese")
x <- "\377"
Encoding(x) <- "latin1"
x
#> [1] "ÿ" 

y <- "\U4E2d"  # Using \Uxxxx always is encoded as UTF-8
y
#> [1] "中"

xy <- paste0(x, y)
xy
#> [1] "ÿ中"
Encoding(xy)
#> [1] "UTF-8"


xy <- paste0(c(x, y), collapse = "")
xy
#> [1] "<ff>中"
Encoding(xy)
#> [1] "UTF-8"
```