@jcheng5 2019-07-24
It somehow didn't sink in for me yesterday that you collapsed WSTextWriter and TextWriter into a single object. I would prefer that they be kept separate; TW presents a simple interface but has a tricky implementation, while WSTW presents a difficult-to-grok interface but has an extremely simple implementation. I found that splitting the respective logic into two separate classes made it easier to prove to myself that each one was individually correct, and that the composition of the two was also correct. With them combined into a single object now, it's harder for me to reason about.

Similarly, I think the logic in writeImpl that collapses the existing buffer should be made into a separate closure; again, I think having smaller parts that do one thing will be easier to reason about in isolation and in combination.

If you disagree let me know, I'm happy to chat about it.

Sorry for not catching this earlier, I'm sure you were expecting this to be ready to merge.