@trestletech 2019-07-24
I agree that the buffer collapsing stuff should be pulled into its own closure. I'll do that!

I totally agree that the readability went way down when we collapsed them. The motivation was just performance to avoid the extra dispatch. I left the breadcrumbs over [here](https://github.com/rstudio/htmltools/pull/131#issuecomment-510141535) but we saw a 13.8% speedup when I collapsed the two writers into one. To me, getting the performance back to on-par with master was worth the readability hit, but if you disagree I'm happy to pull it back out.